refactor: extract SongPreview component
