+++ src/components/host-aware/song-preview.css
@@ -1,0 +1,11 @@
+.song-preview {
+  padding: 0 32px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 38px;
+
+  .song-preview-item {
+    font-size: 12px;
+  }
+}
